revert changes to ParticipationBehind
